Add optional minimum data length parameter to device JSON.
When defining device detection conditions in some cases it may be required to check the length of
the service or manufacturer data on a per device basis.

This change allows for an optional value to be added to the "conditions" array for the minimum data length to be specified.

Before:
"condition":["manufacturerdata", "index", 0, "4c00"]

After:
"condition":["manufacturerdata", 50, "index", 0, "4c00"] // check that the manufacturer data is at least 50 bytes in length.